Basic_Model_Run

Run basic version of SEIRV model

Usage

Basic_Model_Run(
  FOI_spillover = 0,
  R0 = 1,
  vacc_data = list(),
  pop_data = list(),
  year0 = 1940,
  mode_start = 0,
  n_particles = 1,
  n_threads = 1,
  year_end = 2000,
  year_data_begin = 1999,
  vaccine_efficacy = 1,
  start_SEIRV = list(),
  dt = 1
)

Arguments

FOI_spillover

Force of infection due to spillover from sylvatic reservoir

R0

Reproduction number for urban spread of infection

vacc_data

Vaccination coverage in each age group by year

pop_data

Population in each age group by year

year0

First year in population/vaccination data

mode_start

Flag indicating how to set initial population immunity level in addition to vaccination If mode_start=0, only vaccinated individuals If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity If mode_start=2, use SEIRV input in list from previous run(s)

n_particles

number of particles to use

n_threads

number of threads to use

year_end

year to run up to

year_data_begin

year to begin saving data

vaccine_efficacy

Proportional vaccine efficacy

start_SEIRV

SEIRV data from end of a previous run to use as input

dt

Time increment in days to use in model (should be either 1.0 or 5.0 days) '

Details

Accepts epidemiological + population parameters and model settings; runs basic version of SEIRV model for one region over a specified time period for a number of particles/threads and outputs time-dependent SEIRV values. (Differs from Full_Model_Run in not recording new infection numbers or total force of infection.)