Delay_Model_Run_Many_Reps

Run delay version of SEIRV model for single region for large number of repetitions

Usage

Delay_Model_Run_Many_Reps(
  FOI_spillover = 0,
  R0 = 1,
  vacc_data = list(),
  pop_data = list(),
  years_data = c(1940:1941),
  start_SEIRV = list(),
  output_type = "full",
  year0 = 1940,
  mode_start = 0,
  vaccine_efficacy = 1,
  time_inc = 1,
  n_reps = 1,
  division = 10
)

Arguments

FOI_spillover

Force of infection due to spillover from sylvatic reservoir

R0

Basic reproduction number for urban spread of infection

vacc_data

Projected vaccination-based immunity (assuming vaccine_efficacy=1) by age group and year

pop_data

Population by age group and year

years_data

Incremental vector of years denoting years for which to save data

start_SEIRV

SEIRV data (including E_delay and I_delay) from end of a previous run to use as input

output_type

Type of data to output: "full" = SEIRVC + FOI for all steps and ages "case" = annual total new infections (C) summed across all ages "sero" = annual SEIRV summed across all ages "case+sero" = annual SEIRVC summed across all ages "case_alt" = annual total new infections not combined by age "case_alt2" = total new infections combined by age for all steps

year0

First year in population/vaccination data

mode_start

Flag indicating how to set initial population immunity level in addition to vaccination If mode_start=0, only vaccinated individuals If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity If mode_start=2, use SEIRV input in list from previous run(s)

vaccine_efficacy

Proportional vaccine efficacy

time_inc

Time increment in days to use in model (should be 1.0, 2.5 or 5.0 days)

n_reps

Number of repetitions (used to set number of particles and threads)

division

Number of particles to run in one go (up to 20) '

Details

Accepts epidemiological + population parameters and model settings; runs SEIRV model for one region over a specified time period for a number of repetitions and outputs time-dependent SEIRV values, infection numbers and/or total force of infection values. Variation of Model_Run_Delay() used for running a large number of repetitions (>20).