MCMC2

Combined MCMC Multi-Region - series of MCMC steps for one or more regions

Usage

MCMC2(
  pars_ini = c(),
  input_data = list(),
  obs_sero_data = NULL,
  obs_case_data = NULL,
  obs_outbreak_data = NULL,
  Niter = 1,
  type = NULL,
  pars_min = NULL,
  pars_max = NULL,
  n_reps = 1,
  mode_start = 0,
  prior_type = "zero",
  dt = 1,
  enviro_data = NULL,
  R0_fixed_values = NULL,
  vaccine_efficacy = NULL,
  p_rep_severe = NULL,
  p_rep_death = NULL,
  filename_prefix = "Chain"
)

Arguments

pars_ini

Initial values of parameters to be fitted. These should always be the log() values of the actual epidemiological parameters, ordered as follows: 1) Parameters controlling the value of spillover force of infection FOI, either a) a number of FOI values equal to the total number of regions to be considered or b) a number of environmental coefficients used to calculate FOI values from environmental covariates equal to the number of environmental covariates listed in the enviro_data frame. Values should be in alphabetical order by region in case (a) or in the order of the columns in the environmental data frame in case (b). 2) If the basic reproduction number for human-human transmission R0 is to be fitted (i.e. type is set to "FOI+R0" or "FOI+R0 enviro"), parameters controlling the value of R0, either a) a number of R0 values equal to the total number of regions to be considered or b) a number of environmental coefficients used to calculate R0 values from environmental covariates equal to the number of environmental covariates listed in the enviro_data frame. Values should be in alphabetical order by region in case (a) or in the order of the columns in the environmental data frame in case (b). 3) Values of the additional parameters (vaccine efficacy vaccine_efficacy, severe case reporting probability p_rep_severe and fatal case reporting probability p_rep_death) if these are to be fitted, in the order vaccine_efficacy->p_rep_severe->p_rep_death. If these parameters are to be fitted, the values separately supplied to this function (see below) should be set to NULL, the default.

input_data

List of population and vaccination data for multiple regions (created using data input creation code and usually loaded from RDS file)

obs_sero_data

Seroprevalence data for comparison, by region, year & age group, in format no. samples/no. positives

obs_case_data

Annual reported case/death data for comparison, by region and year, in format no. cases/no. deaths

obs_outbreak_data

Outbreak Y/N data for comparison, by region and year, in format 0 = no outbreaks, 1 = 1 or more outbreak(s)

Niter

Total number of steps to run

type

Type of parameter set (FOI only, FOI+R0, FOI and/or R0 coefficients associated with environmental covariates); choose from "FOI","FOI+R0","FOI enviro","FOI+R0 enviro"

pars_min

Lower limits of parameter values if specified

pars_max

Upper limits of parameter values if specified

n_reps

Number of times to repeat calculations to get average likelihood at each step

mode_start

Flag indicating how to set initial population immunity level in addition to vaccination If mode_start = 0, only vaccinated individuals If mode_start = 1, shift some non-vaccinated individuals into recovered to give herd immunity

prior_type

Text indicating which type of calculation to use for prior probability If prior_type = "zero", prior probability is always zero If prior_type = "flat", prior probability is zero if FOI/R0 in designated ranges, -Inf otherwise If prior_type = "exp", prior probability is given by dexp calculation on FOI/R0 values If prior_type = "norm", prior probability is given by dnorm calculation on parameter values

dt

time increment in days (must be 1 or 5)

enviro_data

Data frame containing values of environmental covariates; set to NULL if not in use

R0_fixed_values

Values of R0 to use if only FOI is subject to fitting (i.e. type set to "FOI" or "FOI enviro"); set to NULL if not in use

vaccine_efficacy

Vaccine efficacy (set to NULL if being varied as a parameter)

p_rep_severe

Probability of observation of severe infection (set to NULL if being varied as a parameter)

p_rep_death

Probability of observation of death (set to NULL if being varied as a parameter)

filename_prefix

Prefix of names for output files '

Details

This is the master function for running a Markov chain to optimize the parameters of the yellow fever model based on the calculated likelihood of observing supplied data given a particular set of parameters. [TBC]