Model_Run_Delay_Reactive

Runs delay+reactive version of SEIRV model

Usage

Model_Run_Delay_Reactive(
  FOI_spillover = 0,
  R0 = 1,
  vacc_data = list(),
  pop_data = list(),
  years_data = c(1940:1941),
  start_SEIRV = list(),
  year0 = 1940,
  mode_start = 0,
  vaccine_efficacy = 1,
  time_inc = 1,
  n_particles = 1,
  n_threads = 1,
  deterministic = FALSE,
  response_delay = 56,
  p_rep = c(0, 0),
  case_threshold = Inf,
  cluster_threshold = Inf,
  vacc_cov_cam = c(),
  t_cam = 0
)

Arguments

FOI_spillover

Force of infection due to spillover from sylvatic reservoir

R0

Basic reproduction number for urban spread of infection

vacc_data

Projected vaccination-based immunity (assuming vaccine_efficacy=1 and in absence of emergency campaign) by age group and year

pop_data

Population by age group and year

years_data

Incremental vector of years denoting years for which to save data

start_SEIRV

SEIRV data (including E_delay and I_delay) from end of a previous run to use as input

year0

First year in population/vaccination data

mode_start

Flag indicating how to set initial population immunity level in addition to vaccination If mode_start=0, only vaccinated individuals If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity If mode_start=2, use SEIRV input in list from previous run(s)

vaccine_efficacy

Proportional vaccine efficacy

time_inc

Time increment in days to use in model (should be 1.0, 2.5 or 5.0 days)

n_particles

number of particles to use

n_threads

number of threads to use

deterministic

TRUE/FALSE - set model to run in deterministic mode if TRUE

response_delay

Delay time in days between a threshold being reached and emergency conditions coming into effect

p_rep

Probabilities of an infection being reported as a case before emergency conditions triggered (1st value) or after emergency conditions triggered (2nd value)

case_threshold

Threshold total no. reported cases to trigger emergency conditions

cluster_threshold

Threshold current infectious fraction to trigger emergency conditions

vacc_cov_cam

Target vaccination coverage by age group during emergency campaign

t_cam

Duration in days of emergency vaccination campaign '

Details

Accepts epidemiological + population parameters and model settings; runs delay/reactive SEIRV model for one region over a specified time period for a number of particles/threads and outputs time-dependent SEIRV values, infection numbers and total force of infection values. This version of the model differs from the standard one in simulating an emergency vaccination campaign applied when an outbreak is declared (as well as using delay instead of rate for incubation, infectious period etc.). Case reporting is governed by an additional parameter p_rep which can also change after a reported outbreak is triggered in order to reflect changes in surveillance. An outbreak is declared when the number of reported cases or the infected fraction of the population exceed supplied thresholds.