Model_Versions_Run

TBA

Usage

Model_Versions_Run(
  version = "Basic",
  FOI_spillover = 0,
  R0 = 1,
  vacc_data = list(),
  pop_data = list(),
  years_data = c(1940:1941),
  year0 = 1940,
  vaccine_efficacy = 1,
  time_inc = 1,
  mode_start = 0,
  start_SEIRV = list(),
  mode_time = 0,
  n_particles = 1,
  n_threads = 1,
  deterministic = FALSE,
  response_delay = 56,
  p_rep = c(0, 0),
  case_threshold = Inf,
  cluster_threshold = Inf,
  vacc_cov_cam = c(),
  t_cam = 0
)

Arguments

version

Model version to use; choose from "Basic", "Delay", "Reactive" or "Split"

FOI_spillover

Vector of values of force of infection due to spillover from sylvatic reservoir (size depends on mode_time)

R0

Vector of values of basic reproduction number for urban spread of infection (size depends on mode_time)

vacc_data

Projected vaccination-based immunity (assuming vaccine_efficacy = 1) by age group and year

pop_data

Population by age group and year

years_data

Incremental vector of years denoting years for which to save data

year0

First year in population/vaccination data

vaccine_efficacy

Proportional vaccine efficacy

time_inc

Time increment in days to use in model (should be 1.0, 2.5 or 5.0 days)

mode_start

Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start = 0, only vaccinated individuals
If mode_start = 1, shift some non-vaccinated individuals into recovered to give herd immunity (stratified by age)
If mode_start = 2, use SEIRV input in list from previous run(s)

start_SEIRV

SEIRV data from end of a previous run to use as input (if mode_start = 2)

mode_time

Type of time dependence of FOI_spillover and R0 to be used:
If mode_time = 0, no time variation (constant values)
If mode_time = 1, FOI/R0 vary annually without seasonality (number of values = number of years to consider)
If mode_time = 2, FOI/R0 vary with monthly seasonality without inter - annual variation (number of values = 12)
If mode_time = 3, FOI/R0 vary with daily seasonality without inter - annual variation (number of values = 365/dt)
If mode_time = 4, FOI/R0 vary annually with monthly seasonality (number of values = 12 * number of years to consider)
If mode_time = 5, FOI/R0 vary annually with daily seasonality (number of values = (365/dt) * number of years to consider)

n_particles

number of particles to use

n_threads

number of threads to use

deterministic

TRUE/FALSE - set model to run in deterministic mode if TRUE

response_delay

(Reactive version only) Delay time in days between a threshold being reached and emergency conditions coming into effect

p_rep

(Reactive version only) Probabilities of an infection being reported as a case before emergency conditions triggered (1st value) or after emergency conditions triggered (2nd value)

case_threshold

(Reactive version only) Threshold total no. reported cases to trigger emergency conditions

cluster_threshold

(Reactive version only) Threshold current infectious fraction to trigger emergency conditions

vacc_cov_cam

(Reactive version only) Target vaccination coverage by age group during emergency campaign

t_cam

(Reactive version only) Duration in days of emergency vaccination campaign '

Details

TBA