get_outbreak_data

Generate outbreak data from model output

Usage

get_outbreak_data(
  case_data = c(),
  years_data = c(),
  p_severe_inf = 0.12,
  p_death_severe_inf = 0.39,
  p_rep_severe = 1,
  p_rep_death = 1,
  max_case_interval = 10,
  flag_additional_output = FALSE
)

Arguments

case_data

Vector of cases by time point summed over age

years_data

Vector of year values corresponding to case data

p_severe_inf

Probability of an infection being severe

p_death_severe_inf

Probability of a severe infection resulting in death

p_rep_severe

Probability of a severe infection being reported

p_rep_death

Probability of a death being reported

max_case_interval

Maximum number of days between reported cases before outbreak is declared over

flag_additional_output

TRUE/FALSE flag indicating whether to output detailed case data

Details

Takes in case (infection) data output by YEP::Model_Run(), uses severe and fatal infection rates and reporting probabilities to calculate the number of reported cases, and outputs a list of reported outbreaks. An outbreak is assumed to be reported when one or more cases is reported, and continues until no new cases have been reported for 10 days.