mcmc_checks2

Perform checks on MCMC inputs

Usage

mcmc_checks2(
  pars_ini = c(),
  n_regions = 1,
  type = NULL,
  pars_min = c(),
  pars_max = c(),
  prior_type = NULL,
  enviro_data = NULL,
  R0_fixed_values = NULL,
  vaccine_efficacy = NULL,
  p_rep_severe = NULL,
  p_rep_death = NULL
)

Arguments

pars_ini

= Initial parameter values

n_regions

= Number of regions

type

Type of parameter set (FOI only, FOI+R0, FOI and/or R0 coefficients associated with environmental covariates); choose from "FOI","FOI+R0","FOI enviro","FOI+R0 enviro"

pars_min

Lower limits of parameter values if specified

pars_max

Upper limits of parameter values if specified

prior_type

Text indicating which type of calculation to use for prior probability If prior_type = "zero", prior probability is always zero If prior_type = "flat", prior probability is zero if FOI/R0 in designated ranges, -Inf otherwise If prior_type = "exp", prior probability is given by dexp calculation on FOI/R0 values If prior_type = "norm", prior probability is given by dnorm calculation on parameter values

enviro_data

Values of environmental covariates (if in use)

R0_fixed_values

Values of R0 to use if not being fitted (set to NULL if R0 is being fitted)

vaccine_efficacy

Vaccine efficacy (set to NULL if being varied as a parameter)

p_rep_severe

Probability of observation of severe infection (set to NULL if being varied as a parameter)

p_rep_death

Probability of observation of death (set to NULL if being varied as a parameter)

Details

This function, which is called by MCMC2(), performs a number of checks on data to be used in fitting to ensure proper functionality. It verifies that the number of parameters being fitted is consistent with other settings, that certain values are not outwith sensible boundaries (e.g. probabilities must be between 0 and 1) and [TBA]