mcmc_prelim_fit2

Test multiple sets of parameters randomly drawn from range between maximum and minimum values in order to find approximate values giving maximum likelihood

Usage

mcmc_prelim_fit2(
  n_iterations = 1,
  n_param_sets = 1,
  n_bounds = 1,
  type = NULL,
  pars_min = NULL,
  pars_max = NULL,
  input_data = list(),
  obs_sero_data = list(),
  obs_case_data = list(),
  obs_outbreak_data = list(),
  n_reps = 1,
  mode_start = 0,
  prior_type = "zero",
  dt = 1,
  enviro_data = NULL,
  R0_fixed_values = c(),
  vaccine_efficacy = NULL,
  p_rep_severe = NULL,
  p_rep_death = NULL
)

Arguments

n_iterations

= Number of times to run and adjust maximum/minimum

n_param_sets

= Number of parameter sets to run in each iteration

n_bounds

= Number of parameter sets (with highest likelihood values) to take at each iteration to create new maximum/minimum values

type

Type of parameter set (FOI only, FOI+R0, FOI and/or R0 coefficients associated with environmental covariates); choose from "FOI","FOI+R0","FOI enviro","FOI+R0 enviro"

pars_min

Initial lower limits of parameter values

pars_max

Initial upper limits of parameter values

input_data

List of population and vaccination data for multiple regions (created using data input creation code and usually loaded from RDS file)

obs_sero_data

Seroprevalence data for comparison, by region, year & age group, in format no. samples/no. positives

obs_case_data

Annual reported case/death data for comparison, by region and year, in format no. cases/no. deaths

obs_outbreak_data

Outbreak Y/N data for comparison, by region and year, in format 0 = no outbreaks, 1 = 1 or more outbreak(s)

n_reps

Number of times to repeat calculations to get average likelihood at each step

mode_start

Flag indicating how to set initial population immunity level in addition to vaccination If mode_start = 0, only vaccinated individuals If mode_start = 1, shift some non-vaccinated individuals into recovered to give herd immunity

prior_type

Text indicating which type of calculation to use for prior probability If prior_type = "zero", prior probability is always zero If prior_type = "flat", prior probability is zero if FOI/R0 in designated ranges, -Inf otherwise If prior_type = "exp", prior probability is given by dexp calculation on FOI/R0 values If prior_type = "norm", prior probability is given by dnorm calculation on parameter values

dt

time increment in days (must be 1 or 5)

enviro_data

Values of environmental variables (if in use)

R0_fixed_values

Values of R0 to use if not being fitted

vaccine_efficacy

Vaccine efficacy (set to NULL if being varied as a parameter)

p_rep_severe

Probability of observation of severe infection (set to NULL if being varied as a parameter)

p_rep_death

Probability of observation of death (set to NULL if being varied as a parameter) '

Details

This function is used to estimate the model parameter values giving maximum likelihood; it is primarily intended to be used to generate initial parameter values for Markov Chain Monte Carlo fitting (using the mcmc() function). [TBC]